backend/api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
sqlalchemy==2.0.36
pymysql==1.1.1
python-dotenv==1.0.1
openai==1.54.3